+++ packages/frontend/src/components/common/Filters/FilterGroupForm.tsx
@@ -185,7 +185,7 @@ const FilterGroupForm: FC<Props> = ({
                     leftIcon={<MantineIcon icon={IconPlus} />}
                     onClick={onAddFilterRule}
                 >
-                    Add filter
+                    Add group rule
                 </Button>
             )}
         </FilterGroupWrapper>