+++ packages/frontend/src/hooks/useSavedQuery.ts
@@ -187,6 +187,7 @@ export const useSavedQueryDeleteMutation = () => {
         onSuccess: async () => {
             await queryClient.invalidateQueries('spaces');
             await queryClient.invalidateQueries('space');
+            await queryClient.invalidateQueries('pinned_items');
             await queryClient.invalidateQueries(
                 'most-popular-and-recently-updated',
             );